Projection Method - Centered scheme for convective term
This is the first version of the NS-Python solver implemented using the lectures of Gretar Trygvassson (https://www3.nd.edu/~gtryggva/CFD-Course/).

It considers a centred scheme for convective terms.

The parameters and run options are read by the model through a control file called control.dat. The model is programmed using object-oriented programming (OOP). The parameters, flow variables and grid are objects with their respective initialisation methods.

The model is capable to save outputs in ascii and tecplot format.